% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot_custom.R
\name{pca_plot_custom}
\alias{pca_plot_custom}
\title{PCA plot}
\usage{
pca_plot_custom(pc, comps, nc1, nc2, colorby, shapeby = NULL,
  highlight = NULL, textby = NULL, axisLimits = NULL)
}
\description{
This function is a pca plot with .
}
\examples{
pca_plot_custom()
}
\keyword{pca_plot}
